# Refactor internal getvalue for JSON3.Array to improve performance