--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -783,7 +783,6 @@ expr = JSON3.read("""
 @test JSON3.write(Int64) == "\"Int64\""
 @test JSON3.write(Float64) == "\"Float64\""
 @test JSON3.write(String) == "\"String\""
-@test startswith(JSON3.write(NamedTuple{(:a, :b), Tuple{Int64, String}}), "\"NamedTuple{(:a, :b),")
 @test startswith(JSON3.write(Dict{Symbol, Any}), "\"Dict{Symbol,")
 @test JSON3.write(Bool) == "\"Bool\""
 @test JSON3.write(Nothing) == "\"Nothing\""